app: refactor ValidateMembership to use offline synced space ids (#6371)
`ValidateMembership` checks user membership and shows various screens
during startup. I have a ticket to make the join flow not flash between
the skeleton view and the public town page once you login. This is in
prep for that.

Loading space data and space membership takes too long to smooth out
these various screens. So I added an entry to `use-offline-store` to
store any space ids you've synced. Any thoughts about storing this
elsewhere? The entry will be updated when `useSpaceDataStore` sets
state, which is when a user has synced a space.

So `ValidateMembership` uses this entry as a way to determine whether to
show the user a skeleton or the public town page _before_ the space has
synced. The file is a bit confusing, I rearranged it a bit so hopefully
its simpler. I believe I tried deleting the on chain space check
previously and messed something up, but navigating around the app with
this seems like nothing is broken...