app: use stored data when logging in on Public Town Page (#6390)
This won't make any noticeable difference yet.
For the transition between logged out to in on the public town page, I
want to wrap some of the routes with a Suspense that loads the public
town page (not in this PR).
I found that doing so, `useParams` doesn't match the `spaceSlug` for a
few renders, which results in an undefined space id and a miss in the
the cached space info data we've already stored in the
`useOfflineStore`. Swapping to `useSpaceIdFromPathname` the spaceId is
always defined.