Tak/hnt 5874 notification text not decrypting because of cryptostore database name now has extra suffix (#6428)
The CryptoStore name now has extra suffix that looks like this:

```typescript
// In TownsClient.ts
const envSuffix = this.opts.environmentId === 'testnet' ? '' : `-${this.opts.environmentId}`
const cryptoDbName = `database-${userId}${envSuffix}`
```

The ServiceWorker has no access to the dbname directly. Using existing
sync between the client and the SW, store the crypto db name in
the`NotificationCurrentUser` store. When a new push event comes in, the
SW will use the NSCU store to get the db name, and then use it to open
the correct `CryptoStore` to get session keys for decrypting the
cipertext.