app: store unauthenticated data in session storage (#6421)
moved useFetchUnauthenticatedActivity to its own file and changed from
react state to zustand session storage, and made requests abortable

For transitioning from logged out to in on public town page, we have to
remount the public town component, so even without flashing other
screens the content gets re-rendered. This will at least pull data from
session storage while doing this.

It just plops/updates the data in the store as it comes in, and doesn't
prevent refetch after transitioning to logged in state, so you'll
actually see the data cycle repeat itself but at least its not a huge
jarring content render. We can optimize this later if we offload data
retrieval to some service.