Kerem/blue green (#6540)
* enables standby mode
* updates the healthcheck endpoint to /status
* sets deployment minimum and maximum percentages to allow blue-green

@jterzis i'm tagging you here because we have new environment variables
we'll need to put in documentation. namely, `STANDBYONSTART`

Note: transitioning from the previous configuration where supervisord
autorestarted the stream node vs this one where it doesn't, will require
us to first stop the currently running nodes, and then start the new
batch. because blue will still be on restart mode when we run this infra
migration. but further deployments won't have this issue.