Tak/hnt 5685 fix notification navigation for mobile pwa not routing to the right path (#6546)
Issue is that the bookmark code has not been updated to work with DM /
GDM and notification. Current bookmark stores the spaceId and channelId
as the user taps on the phone. When a notification pops up, and the user
taps on it, if it is a DM / GDM, the app is supposed to navigate to
those channels. However, the bookmark code is not updated to deal with
that. Instead, it navigates to the bookmarked spaceId/channelId path.
Consequently, when a user taps on a notification, it appears that the
app navigated to some random place.

Fix:
- refactor and consolidate the routing code in `SpaceContextRoute.ts`.
Keep the logic for figuring out the routes for desktop and mobile in one
place.
- log the url searchParams to DataDog. Will
delete the logs after confirming that the routing logic works.