lib: fix bad check in TownsClient.joinTown (#6560)
When we split the repo and updated all the tests for xchain I either got
confused and put this insane check in place, or else something has
changed in the last two months where all the tests in joinSpace now
pass.

To note, the joinSpace tests (and other tests) still do NOT test gated
spaces. When they do, these tests do not all pass.