UserOps: retry when preverification gas fails (#6735)
This adds retry logic to user operations specifically when we get
preverification gas errors.
https://www.stackup.sh/blog/an-analysis-of-preverificationgas

When the op is sponsored we cannot estimate the gas, the paymaster does
its own gas estimates for the operation. There is still a chance that
this estimate could be too low, resulting in a failed op. The best we
can do in this case is retry, for sponsored ops this will retry 3x
default.

For non sponsored ops (i.e joining a paid town, or any other op rejected
by the paymaster/worker). In this case, we'll bump the estimate by 10%
each retry until the user dismisses the confirmation.

cc @erikolsson in case any of this is relevant for iOS