App: add header to unjoined channels component and fix member count for channels the user is not part of (#6758)
sorry for the big pr folks.

This pull request adds a header to the unjoined channels component and
fixes the member count for channels that the user is not a part of. The
changes include adding a new file, `useUnjoinedChannelMembers.ts`, which
provides a function to get members from unjoined rooms and caches the
results. Additionally, the `ChannelUsersPill` component has been updated
to use the `useChannelHeaderMembers` hook instead of `useChannelMembers`
to retrieve member IDs from joined/unjoined channels. The
`ChannelHeader` component has also been updated to use the
`useChannelHeaderMembers` hook and display a loading indicator only if
the user is a member of the channel. Finally, the `TouchChannelHeader`
component has been updated to use the `useChannelHeaderMembers` hook and
display the correct member count for unjoined channels.


![image](https://github.com/HereNotThere/harmony/assets/157021769/9b601673-1ce9-4201-a911-a8adae48e431)