notification deep linking for @mention and @channel in a reply (#6838)
In a reply, if the user @mention or @channel, then tapping on the
notification should deep link to the thread. Add the threadId from
client -> notification-service -> serviceWorker.
In a thread, @channel is a special case because the incoming
notification will appear as a new message for everyone in the channel,
and not just for the thread participants. Add notification parsing and
handling for everyone else in the channel who is not a thread
participant.