app: improvements to feed resizing (#6913)
some bits discovered while looking at sending attachments. Since the
attachments take more space in the rich-text area, sending them makes
the v-list container shrink uncovering a couple of old bugs:
- if the available scroll-area is less than the size difference resulted
by the resizing of the container, an implicit scroll will get triggered
(to fit the contents within the scrollable area) in addition to the
correction. This offsets the scroll area by the latter.
- useLayoutEffects to make sure the offsets are computed while hitting
the screen

also updated the examples in playground