fixing turbo cache (#6946)
- Reintroducing "dist/**" to turbo cache
- Removing the "maintain turbo cache" step from the multinode job,
because this job skips cache reads during turbo build.
- Updating cache key to use the branch name to maintain cache coherence
between multiple runs. `run-id` will change in each run, but stay the
same within re-runs. that means we're missing out on performance gains.
- Nit fix: Updating restore-keys to fallback to cache state from the
main branch. (restore-keys is what turbo uses as a fallback when the
actual cache `key` returns a cache miss). If we don't change this key,
the first time a CI job runs, we'll be picking up the cache from a
completely arbitrary workflow run. This restored cache could be coming
from a separate PR for example, which is not too terrible but cache
state from `main` is likely to be closer to the current PR, which is a
slightly more performant, and a more deterministic option.