Render Build - Slack Notifications Pt2 (#6957)
This PR enables the render_build job to run on merge to main and omega.
It fixes the environment variables, as well as a few bugs in the
deployment script. It also treats cancelled deployments as success,
since each new commit will cancel an ongoing deployment - which should
not trigger a failure notification.