app: enable sending messages while uploading - take 2 (hnt-4902) (#6978)
take 2, mimicking Slack

The event order will not reflect the order of interaction but rather the
order of settlement (as opposed to [this
approach](https://github.com/HereNotThere/harmony/pull/6900)).
Initially, I felt this was buggy and abandoned the idea. However, after
seeing that Slack does exactly this, I started thinking it also has pros
in terms of UX.

The icing on the cake would be to settle the order on the timeline based
on the one from the server once the process is done, currently this only
happens on reload. Sorting the events by `confirmedEventNum` fixes this
but you may have a better idea @texuf


https://github.com/HereNotThere/harmony/assets/557990/382ca354-0901-47d3-93cc-3d6616ac47f8




- push event
  - retrieve local id
  - store references to file storage to follow progress
- upload files 
  - map progress to timeline 
- all attachments uploaded:
  - transform attachments, amend local event
  - let go remote event