editor: clicking on list format button does not format all lines [HNT-6611] (#7040)
This was quite the bender. When we click on list icon, Plate formats all
paragraphs into `<li>`. However, in our editor, `shift+enter` creates a
new line, and not a new paragraph. Had to implement this function to
break new lines into new paragraphs, while maintaining other
non-selected text and the selection itself.

---------

Signed-off-by: Abhinav Pandey <abhinav@hntlabs.com>