omega releases will trigger archive node deployments (#7100)
pull_river.sh deploys stream nodes and archive nodes to gamma. and this
pr takes care of the same thing for omega. i think this is the most
straightforward way to keep them up to date. whenever we execute a
release on omega, this script will pull the latest "mainnet" docker tag
and redeploy it to omega archive nodes