app: name confirm prompt + fix channel names overriding (hnt-6895) (#7127)
Previously a space user could in theory inherit info from a channel user
if the username `!== ''` this would never occur since space channel
usernames are always `''` because we never set them, we only do so for
GDMs.

What broke in the previous (reverted) PR:
I relaxed the `meaningfulInfo()` check to to check for not `undefined`
instead of `''`. This is necessary for spaces to detect an empty user
and trigger the confirm name prompt. A side effect of this change was
that it triggered the aforementioned bug.

This PR separates `setSpaceUser` and `setChannelUser` to avoid confusion
(space user will never inherit from channel user anyways) - this way we
can have a more relaxed check for spaces, where we even allow users
without usernames so we can prompt to set it, and in the case of GDMs,
we will only allow the ones that contain a set displayName.

I also removed the setUser per channel (space channel, not GDM channel)
since we don't allow setting username per space channel.