Kerem/on chain metrics pt3 (#7098)
this pr extends our on-chain metrics collection to include the
following:

1) nodes on base (via EntitlementChecker)
2) operator status on base
3) stream count on node
4) aggregate network metrics such as: num total spaces, num total nodes
on base, num total nodes on river, num total operators on base

We have a handful of river-build changes that need to make it to include
things like total delegated. You'll also notice that we have a few
additional abis here. I will remove them once we merge this one:
https://github.com/river-build/river/pull/151

This one needs to merge for river operator metrics:
https://github.com/river-build/river/pull/159
This one needed for a more efficient data-fetch on # total streams per
node: https://github.com/river-build/river/pull/147
This one for connecting nodes to operators on base:
https://github.com/river-build/river/pull/158
This one for total delegation:
https://github.com/river-build/river/pull/157