app: disable wallet link toasts only on public town page (#7206)
if the route is a town, but you're not a member, it indicates you are on
public town page, which is when we do not need these toasts.
there are various routes in the app that allow for wallet linking when
not in the town (b/c of route/panel refactors), such as the create page,
DMs, etc. So the toasts were not appearing and the list was not updating
in those cases either.