lib: catch river errors in waitForCreateChannelTransaction (#7221)
I was able to recreate the behavior shown in the video here by just
throwing an error from TownsClient.createChannelRoom.
https://linear.app/hnt-labs/issue/HNT-7172/[omega][user-feedback]-im-going-through-the-process-but-it-doesnt.
So this doesn't solve the issue, I don't know what Billy was running
into exactly, it seems entitlement related, but at least the error will
bubble up to the app.

We track blockchain tx state but not river success/error state and the
hooks used in the client are all wrapped in useWithCatch so the error
was getting swallowed. This is a quick fix, but I think in the future we
should examine all our txs and the various river actions they kick off.
The form in app in the video also deserves a solid refactor but its not
at the top of my todos.