app: add reactive user lookup hooks + implementation examples (#7239)
Adds reactive hooks to retrieve usernames. 

The main one being `useUserLookup(userId:string)` - the hook extracts
the user slice from zustand within the scope of its channel and space.
We should transition to this whenever possible.

For use-cases where we require a list of users to build something as a
whole (drop down search results, list of names,...) I have added
`useUserLookupArray(userId:string[])` and
`useUserLookupMap(userId:string[])`

I think without something like this we won't fully take advantage of the
username refactor.

### Background

When refactoring the user lookup we added an imperative handle
(`lookupUser()`) that enabled us to transition from the previous
implementation to the new one without changing too much logic.

While this works `lookupUser()` will only update the values imperatively
when a component renders. This often works as a happy accident since
user changes are often related to stream updates that triggers component
refreshes from above on the hierarchy. However, the goal is to not rely
on app refreshes to propagate user changes.

### Implementation examples

@brianathere I focused on the scenario you described yesterday (you
invite a user joins you want to @) and you can see how I updated the
code to make it work.
In some places we still need an array or a map with all the usernames, I
added some comments in the code related to that.


https://github.com/HereNotThere/harmony/assets/557990/758af980-548e-447c-8471-b4b2dd7ab9d1

### TODO

I want to rename the hook that provides the imperative lookupUser too
but it's too much noise for this PR
I think we can keep `lookupUser` as we transitions them one by one to
reactive ones. There's also some work to do in places where the username
is stored and passed around (Channel titles) you can see in the video
that the whole sidebar flashes upon update. In general a lot of this is
low hanging fruits .