Kerem/use shims for metrics (#7267)
this pr updates the on-chain-metrics lambda to use `@river-build/web3` +
ethers instead of viem + @river-build/generated`

@giuseppecrj for some reason, `getStreamCountOnNode` on gamma river
registry is still not working. this part was already commented out in
the previous version of the lambda, so i'm still momentarily keeping it
commented out on this iteration. we should figure out why this simple
view function keeps reverting