Kerem/on chain metrics pt3 (#7296)
This PR improves the on chain metrics lambda in a few different ways:

1) Introducing the MetricsExtractor and MetricsIntegrator classes to
better organize the code and prepare it for unit tests
2) Promise.all calls added to improve performance
3) Making use of the new BaseRegistry.getOperators() view function to
remove false entries
4) 2 Additional integrator functions `enrichNodesWithOperators` +
`enrichOperatorsWithNodes`. These functions produce the ultimate view of
nodes and operators on both registries. The former does it from the node
perspective, and the latter does it from the operator perspective. These
allow us to see how every single node on both chains is related to every
single operator on both chains. And vice-versa. We can then pass this
into a json table visualizer to see the state of the registries.

## Screenshot

![image](https://github.com/HereNotThere/harmony/assets/5973785/40f88f3d-bac8-42e0-8e75-a7f311f17979)