add local 4337 servers (#7311)
a modified version of https://github.com/stackup-wallet/erc-4337-devnet

Initially, could not get the chain id in local geth to match 31337 -
1337 is hardcoded in go-ethereum source code for --dev mode.

Stackup also builds geth with custom tracers, and the local paymaster
does not work without them. So I built the client as described in
https://github.com/stackup-wallet/erc-4337-execution-clients with a
chain id to match anvil, and published a docker image, which is what is
used instead of the original stackup image. The end result is a geth
node with chain id of 31337, with stackup's tracers.

This will be run during local development in a dedicated dev flow, part
of another PR.