app(set-ens): revert UX changes, still keeping optimistic update (#7304)
Reverting the optimistic update on select a new ens name -- the DX is
bad, the UX isn't that great.
Added back the button to set the ens: it will optimistically update your
ens name and if it fails, a toast will pop saying that it failed. The
button create a better UX for both mobile and desktop users.

`onClose` doesn't perform the optimistic update anymore.