app: enable unfurling links after send (hnt-6637) (#7405)
- display preview placeholders in editor while fetching previews
(spinner + link + close)
- if when sending there are previews that aren't ready we insert the
message locally during a grace period (10s)
- if the prevews are loaded during this period, the placeholder
attachment payloads are replaced, otherwise the failed attachment get
discarded before the message is sent to network

This is the first iteration. it's seems to work well in terms of UX but
I want to re-iterate the code to enable re-using the former useQuery
(the one that didn't finish before sending) - this doesn't seem to be
possible to achieve in the context of an imperative function call. In
order to make this happen we need to do something similar to the
attachments, i.e. having a unfurl state context which persists.



https://github.com/user-attachments/assets/b8979c7b-c21e-4111-b34e-a42f39fef5a5