stackup-worker: change to privy auth token verification, and optimizations (#7423)
- the client will pass a privy access token when making requests to
stackup-worker
- the worker will verify the token for all requests. if we want to
change this to specific user operations we can, but the 2 ops where we
would put it would be specifically for new user flow anyway, so i think
its ok to use it for all op types.
- removed Privy /user lookup, which was taking 6+ seconds. For checking
email whitelist on create space operations, if strict limits are turned
on (it's not currently), then it will use the /user/privyDid lookup
- removed all verification checks (logFilters) for
`createSpace_linkWallet` and `joinSpace_linkWallet` because they're
specific to new users
- there were various places in the stackup-worker code that
createdSpaceDapp, and then used the spaceDapp to check for things like
if the town exists. I removed a bunch of these because we already do a
bunch of checks in the client like this. Some checks (like
spaceDapp.getMembershipPrice for joinFlow) are retained.

There's no way to test this on Omega (the only place where the privy
query was taking a long time) until we do another harmony deployment
(app + worker). Once in, I can verify if the privy auth token check is
going to meet our needs timewise. If not, Pat and I discussed removing
Privy checks altogether.