userops: cache builder, move middleware (#7469)
caching the userop.js `builder` and `client` in the UserOperations class
can save 1-2s on user operations.
Calling `userops.setup()` will initialize these values.

https://github.com/HereNotThere/harmony/pull/7434 revealed that by
caching the builder, redundant middleware was added with each new
useroperation. Prior, the middleware was reset between each because
every call created a new builder.

So I moved middleware to builder initialization, and added the
MiddlewareVars class for storing various properties used in the
middlewares.

I also added a userops.retryGas.test to ensure the preverification gas
multiplier runs as expected.
With Alchemy, we'll be able to pass this multiplier on to the paymaster.
We could even default the multiplier to higher than 1 if we desire.