app: enable force updates (HNT-9054) (#7500)
### tl;dr

if a breaking change comes we should bump the minor or major version of
the app package.json, e.g. `1.2.3` -> `1.3.0` this would result in
skipping the update prompt (or the lack of the latter in certain areas
of the app).

### main rules for force-updating:

1. if a minor or major version change is detected in `app/package.json`
2. if the path is safe (landing or public town-page landing, ... we
could add to this or use other rules)
3. if a flag stored in cache doesn’t yet exist, denoting that this
version (i.e this PR) isn't yet running. This can be removed in a few
weeks or so.

### changes:
- server: add `/version` path
- sw: make sure `/version` doesn’t get cached (network only)
- app: compare version whenever a new update lands, force update if any
of the first two rules above matches
- sw: check if flag described in rule 3 exists and update accordingly

### misc: 
use `APP_VERSION` from package in a couple of places add release date to
error prompt