app: add verificationGasLimit to confirm tx modal (#7510)
https://app.towns.com/t/10c87bb04477151743070b45a3426938128896ac5d0000000000000000000000/channels/20c87bb04477151743070b45a3426938128896ac5d36c7eb51c50c25fde36f25#94b4df4a984ca55733ade3fbee701300892a081af0e0392963c4a15342643194

Julija incurred a cost greater than what was displayed in the tx modal.
She noted she had 0.009 ETH, and after tx it went down to 0.00272 ETH, a
difference of ~ 0.00628.
The estimate in the modal displayed 0.00497 ETH.

I double checked her balance at the relevant blocks using
https://sepolia.basescan.org/balancecheck-tool?a=0xF1474363Ba7D7fE8AA81010f8AbfA827FE9dA477.

Verification gas limit was not considered as part of the calculation
presented to the user, and the order of operations was wrong too. The
[tx](https://sepolia.basescan.org/tx/0x02d8d5f049cfc7e9b5b026a758aa6e8c0aa13f877e4b141c476cc348dbb2b94d)
gas was
```
0 | ops.callGasLimit | uint256 | 502958
0 | ops.verificationGasLimit | uint256 | 91551
0 | ops.preVerificationGas | uint256 | 50251
0 | ops.maxFeePerGas | uint256 | 9274441634
``` 
(callGasLimit + verificationGasLimit + preverificationGas) x
maxFeePerGas is about 0.0059. + the 0.0003 membership fee = ~0.0062