app: clean user ens data after wallet unlink (#7512)
This PR does changes in the towns client to emit an event about the
wallet unlinking, so we can listen this event and clean the ens data
from the user in our user store.

> The entry in the stream of their ens name is only a hint, not
authoritative

Now the `useUserLookup` will only return the ens name / address if the
user actually owns the ens. i.e. has the ens wallet linked

Previously, you could set the `ensAddress` to whatever you want if and
we would show it.