Autojoin, + wire up hide user join/leave state changes in towns (#7602)
Clients are autojoined to all auto-joined towns. Setting autojoin and
hideUserJoinLeaveEvents on channel creation and updating is exposed via
the towns client.

autojoin and hideUserJoinLeaveEvents are added as fields on the Room
class, and useCasablancaRooms is modified to update whenever autojoin or
hideUserJoinLeaveEvent updates are emitted.

There are a couple of new E2E tests in harmony that test autojoin
behavior.