Kerem/alpha render cd (#7616)
This PR fixes the render_deploy ci job and adapts it to an additional
`deployment_status` that got introduced with fastify. This workflow has
been disabled for a while. I'm temporarily removing `on push` triggers
before re-enabling this workflow. I'm also adding alpha to this
workflow. I will test this manually once we merge this PR and then add
the `on push` triggers back.