create town and edit space info changes for uploading town's image to the river stream (#7621)
Create town flow changes:
- If an image is selected in the Create town flow, at the end of the
contract transaction, the space contract address is used to derive
the `iv` and the `secretKey`.
- Image is created as a chunked media, and is encrypted with the derived
iv and secretKey
- A new media stream is created. The encrypted space image is uploaded
to the media stream
- A spaceImage event to the space stream. The event containts
the `context`. The `context` is the space contract address. Idea is that
any client can derive the iv and the secretKey from the space contract
address.
- This event is snapshot.

SpaceInfo panel icon update changes:
- derive the iv and the secretKey from the space contract address.
- Encrypt the image with the derived iv and secretKey
- Upload the new image to the media stream
- Add spaceImage event to the space stream

Local development:
- added scripts to start and stop the stream-metadata service,=.
- updated tasks.json  and start_dev.sh to call the start / stop scripts