app: towns-11268 pin messages issues with pinning and editing messages (#7778)
#### first:
I realised I was using the wrong event (not timelineEvent) which enables
richer content. This is because I coded this before the helper in the
lib existed, using the bare sdk.

#### second: 
now using latestEventId as a reference when pinning. It's kind of
working but it's also a bit spooky since the pin will disappear from the
timeline message if the message is edited (we are attributing pins based
on eventId === pin.latestEventId).

Writing this I realise we should probably just update the pin, the same
way as we should delete it alongside the message 😄

but I may have misunderstood this completely, what are your thoughts
@texuf

---------

Signed-off-by: Theo Tillberg <ttillberg@gmail.com>