Userops: use rundler_maxPriorityFeePerGas to calculate gas fees for non-sponsred userops (#7877)
Alchemy [recommends using their own
endpoint](https://docs.alchemy.com/reference/bundler-api-fee-logic) for
calculating user op gas fees, as opposed to eth_maxPriorityFeePerGas.

This is needed only for the case of a non-sponsored user operation. For
sponsored operations, the alchemy_requestPaymasterAndData endpoint is
used, which includes gas fee estimates.

In order to accomplish this, I created a TownsSimpleAccount, which is a
[userop.js builder](https://docs.stackup.sh/docs/useropjs-builder), that
mostly clones the SimpleAccount builder we were already using. It just
removes some of the default middleware that came with it (paymaster, gas
fees).