adjusting the notification service resources (#7934)
This PR addresses 2 things:

a) I had a rightsizing adjustment for the notification service, but the
migration did not go through as I had previously configured terraform to
ignore changes on those fields. I noticed this as I was going through
our aws budget.
b) The numbers I had put in there were too conservative, so I'm bumping
them up a bit, in alignment with what I'm seeing in the metrics.