stackup-worker: rename env vars (#7923)
rename PAYMASTER_ADDRESS to ENTRYPOINT_ADDRESS - this var is defined in
the wrangler.toml file, so it should update automatically on deployment

rename PAYMASTER_RPC_URL to STACKUP_PAYMASTER_RPC_URL - this var is a
secret. I have already added a new entry to each environment, with
identical value. Once this PR is merged and deployed to omega, we can
remove PAYMASTER_RPC_URL

Testing:
- [x] once render preview is up - deploy test-beta stackup-worker from
this branch. Verify the render preview sponsors a tx.
- [x] Redeploy test-beta stackup-worker from main. Verify app.gamma
sponsors a tx.