lib: pass ' ' as uri in TownsClient.updateSpaceInfoTransaction (#7942)
This seems to have been a bug introduced by the Towns Metadata project.
Previously, it seems we did not pass a uri for updating a space. That
was updated in the client
https://github.com/HereNotThere/harmony/pull/7337.
Also, it seems the space uri is also being passed as `undefined` to
TownsClient.updateSpaceTransaction, though it's typed as a string.

~~We don't pass a space uri in town creation - the default uri returned
when we grab space data for updating is "". Which doesn't pass contract
checks. Perhaps related is
https://github.com/river-build/river/pull/586~~ This seems mostly b/c of
7337

This seems incomplete, I'm not sure if we're supposed to be passing a
uri on the client during town creation, or if the uri is being set
incorrectly in the contracts during town creation.

For now we can pass this empty string, I don't see it hurting anything
and others can be unblocked by this bug.