Userops: add getUserOperationReceipt polling (#7939)
- Added TownsUseropClient which is a copy of Userops.js Client, but adds
the getUserOperationReceipt method to the sendUserOperationResponse.
This will poll eth_getUserOperationReceipt for the user op.
- Switched TownsClient useropClient.wait ->
useropClient.getUseroperationReceipt, with updated logging
- Currently we only send specific logs to datadog for the join and
create flows. We will continue doing this, but additionally
getUseroperationReceipt will run for _all_ user operations in bg, and
send metrics to DD. So we should be able to get a good read on how long
user operations are taking us across the board. A userOpHash is included
in both the join/town flows, where applicable, as well as in the logging
for getUseroperationReceipt, so ops can be cross referenced between the
views in DD.