app(space-metadata-image-fetching): fix loading and remove react-query layer  (#7952)
Let the browser do the fetching & caching. Storing the same thing in
react-query was just taking more memory and slowing down everything.

We already have control of resources in the image store, so we dont lose
nothing here.

We dont need to manually create imageObjectUrl using
`URL.createObjectURL`. Lets throw this to the browser and let it handle