app: edit message UI is broken on mobile [TOWNS-11431] (#7957)
Before 
![Screenshot 2024-09-10 at 9 11
59 PM](https://github.com/user-attachments/assets/f3ac766c-2f4a-432a-9497-33d66c2565e1)

After
![Screenshot 2024-09-10 at 9 11
20 PM](https://github.com/user-attachments/assets/08541c37-b564-44a3-a65e-2501697a21a7)

Couple of things wrong with it
1. We currently show the MessageEditor as an absolute positioned div on
top of the usual editor but it's portalled into #root. This causes
issues like
   a. `lightHover` is transparent and shows both editor
b. the height of original editor and message editor is not same and
leads to scroll issues + hidden text in timeline

I've added an ID to original message editor's container and portal the
message editor into that div, ensuring there are no scroll & height
issues

---------

Signed-off-by: Abhinav Pandey <abhinav@hntlabs.com>