updating the node ping metric logic for better visualization (#7864)
I would like to replace the datadog synthetic uptime monitors with
custom monitors built on top of our metrics lambda, because:

a) synthetics cost too much
b) it's hard to get to them to run against the entire network (thus we
only get alerted on our own nodes)
c) they only extract success/failure info

This PR makes a few adjustments to the metrics lambda to help us dissect
ping data, and acts as a stepping stone towards the better monitoring
scheme outlined above. I'm also removing an aggregation computation
`numUnhealthyPings`, because we'll be able to derive this in the
dashboards.