app: prompt privy login if not signed into privy (#7963)
Thanks to @erikolsson for the idea.

Applicable to situations in which user is signed into River, but Privy
session has expired, and they want to do a tx.

Previously, we forced users to logout of River and re login when this
happened. And this was actually broken.

New behavior is, user is prompted w/ a toast that tells them we need to
re-auth them to Privy, they can click to re-auth. If they log in w/ the
account that matches their userId/loggedInWalletAddress, they get a
success toast. If they log in w/ a different account, they get an error
toast.