Fix cludge, don’t reset the sender on membership events (#8047)
I’m working on TOWNS-7347 and i need to touch casablanca timelines,
noticed this was weird

This came in here https://github.com/HereNotThere/harmony/pull/6105 and
i should have caught it then, membership events have an initiator on
them, which may or may not be the current user. Just use the right value
in the ui instead of hacking the event.