TOWNS-7347 Don’t render messages if someone tries to edit or delete your message (#8053)
Logging a warning for now. The result of this should be that some
messages that were accidentally deleted show back up and we get a few
warnings in the console.

Also prevent the client from sending edits or redactions if it’s not
your own event. I thought about putting these checks in the sdk, but it
seemed over prescriptive... we can move them there later