app: create space image upload adjustments (towns-7004) (#8172)
I had to change the form slightly to enable requiring image upload. I
hope this is for the better, less layout jumps etc.

- display errors onSubmit
- make submit button green when all fields are ready, if clicked while
not ready errors will show
- custom render props image uploaded to integrate better with the
interactive token
- hide interactive token when panel is open to avoid shift

- possible regression: we're not pushing panel open automatically if
there's an error in the form. If we need this we could do it based on
the type of error.


https://github.com/user-attachments/assets/a27a3799-1c1b-4446-88ed-eea3cc9a937c