TOWNS-12869 Fix the space upload to use a wrapped towns function instead of the auth’d client (#8210)
This is a nuance, but we have a clientSingleton and a client. They both
point to the same thing, but the client is only defined when the user is
logged in. It’s a weird distinction, but it exists. If we use the
wrapped functions they use the singleton. Another option for this fix
would have been to just use the clientSingleton instead of the client.