app: fix no signerContext on public town page while authenticating (#8186)
### Background

This issue happens only for "logged in" users that are trying to join a
town from the public town page, before a TownsClient is started.

The crux of this issue is use (misuse?) of `isAuthenticated` from
`useCasablancaCredentials`, and a dependence on `signerContext` from
`useTownsContext` during the login flow within `lib`.

Scenario presents when a user has previously logged in and casablanca
credentials are saved to local storage. They boot up the app again -
credentials are immediately read from local storage, so
`useCasablancaCredentials.isAuthenticated` is true. However, during the
bootstrap of the towns client within `useTownsClientListener`,
specifically in this block:

```
function loginWithRetries {
...
 } else {
                console.log('**user authenticated, starting casablanca client')
                useOfflineStore.getState().setSkipIsRegisteredCheck(userId)
                const casablancaClient = await client.startCasablancaClient(context)
                return new ConnectedToRiver(credentials, casablancaClient, context)
            }
}
```

The client encounters an error during `client.startCasablancaClient` and
continues to retry logging in.

So the UI checks `isAuthenticated`, determines a user _is_ able to
proceed with joining a town, shows the "Join" button, and user clicks
it. Meanwhile, login retries are going on, which can take very long (or
never and user is ultimately logged out), and the client listener never
ends up updating `signerContext`, so that it remains undefined while
attempting to join.

### Solution
While bootstrapping the client, the public town page will replace the
"Join" button with "Connecting to River". If it doesn't connect after 10
seconds, an error toast is displayed.

It is a bit weird to think we can have an "authenticated" state while we
do not have a client initialized, and the client uses `isAuthenticated`
in a numerous places. We maybe want to convert these cases to checking
`signerContext` or some similar thing - but I can also see that a quick
localStorage check could allow the app to render/load (from local cache)
faster while the client bootstraps.

For now, considering launch, I prefer to scope this fix strictly to the
login/join button, I don't want to introduce some loading weirdness or
bugs elsewhere that `isAuthenticated` is used.