Add a retry when tryng to start a casablancaClient during new user join/create space (#8240)
ref TOWNS-12900
in the above bug we saw that isEntitled failed for a new user when
trying to create a stream. Instead of going crazy and adding block
numbers to all requests we decided to just retry the request a few times
and see if it resoved.
adds console.errors that should show up in datadog.